% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{OnePotSPROX_Phenotype.Fun}
\alias{OnePotSPROX_Phenotype.Fun}
\title{OnePotSPROX Phenotype Analysis}
\usage{
OnePotSPROX_Phenotype.Fun(
  Expression_data,
  SPROX_Raw,
  SD_cutoff = 2,
  p_cutoff = 0.05,
  TMTplex = 10,
  plot = TRUE,
  xlab = "Z Score",
  ylab = "- log10 (p value)",
  labelcolor = c("grey", "red"),
  alpha = 0.5,
  plottextsize = 12,
  correctedpvalue = FALSE,
  interactiveplot = FALSE
)
}
\arguments{
\item{Expression_data}{Raw ProteomeDiscoverer (v. 2.3) data}

\item{SD_cutoff}{A positive integer (default = 2)}

\item{p_cutoff}{A positive integer between 0 and 1 (default = 0.05)}

\item{TMTplex}{8, 10, 16 TMTplex (default = 10)}

\item{plot}{Display volcano plot (default = TRUE)}

\item{xlab}{X-axis Label (default = "Z Score")}

\item{ylab}{Y-axis Label (default = "- log10 (p value)")}

\item{labelcolor}{Color of Not significant and Significant data points (default = c("grey", "red"))}

\item{alpha}{Transparency of plot (default = 0.5)}

\item{plottextsize}{Text Size of plot (default = 12)}

\item{correctedpvalue}{Bonferroni corrected p-value (default = FALSE)}

\item{interactiveplot}{Plotly interactive Volcano plot (default = FALSE)}

\item{TPP_data}{Raw OnePotSPROX (TMT10plex) ProteomeDiscoverer (v. 2.3) data}
}
\value{
Volcano plot, Hit list, Total proteins assayed and Unique protein hits
}
\description{
This function will take expression data and OnePotSPROX (TMT10plex) data directly from ProteomeDiscoverer (v. 2.3) and generate a volcano plot and hit list
based on user-defined SD and p-values
}
