% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{OnePotTPP.Fun}
\alias{OnePotTPP.Fun}
\title{Load Expression data and TPP data}
\usage{
OnePotTPP.Fun(Expression_data, TPP_Raw, SD_cutoff, p_cutoff)
}
\arguments{
\item{infile}{Path to the expression input file, TPP data, SD_cutoff, and p_cutoff}
}
\value{
Hit list, volcano plot
}
\description{
This function will take expression data and TPP data directly from Proteome Discoverer and generate a hit list
and Volcano Plot
}
