% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{STEPP.Fun}
\alias{STEPP.Fun}
\title{STEPP Analysis}
\usage{
STEPP.Fun(Expression_data, STEPP_Raw, SD_cutoff, p_cutoff)
}
\arguments{
\item{Expression_data}{Raw ProteomeDiscoverer (v. 2.3) data}

\item{SD_cutoff}{A number}

\item{p_cutoff}{A number}

\item{TPP_data}{Raw STEPP (TMT10plex) ProteomeDiscoverer (v. 2.3) data}
}
\value{
Volcano plot, Hit list, Total proteins assayed and Unique protein hits

Hit list, volcano plot
}
\description{
This function will take expression data and STEPP (TMT10plex) data directly from ProteomeDiscoverer (v. 2.3) and generate a volcano plot and hit list
based on user-defined SD and p-values
}
