% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{Two_Comparison_Venn_Diagram.Fun}
\alias{Two_Comparison_Venn_Diagram.Fun}
\title{Two Method Comparison Venn Diagram}
\usage{
Two_Comparison_Venn_Diagram.Fun(data1, data2, name1, name2)
}
\arguments{
\item{infile}{Path to the data1, data2, name1, and name2}
}
\value{
Venn Diagram
}
\description{
This function will make a Venn Diagram showing unique hits for two different proteomics methods
}
