% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{Venn2.Fun}
\alias{Venn2.Fun}
\title{Two Method Comparison Venn Diagram}
\usage{
Venn2.Fun(data1, data2, name1, name2)
}
\arguments{
\item{data1}{Unqiue Hit .csv Output method 1}

\item{data2}{Unique Hit .csv Output method 2}

\item{name1}{Method 1 name}

\item{name2}{Method 2 name}
}
\value{
Venn Diagram
}
\description{
This function will make a Venn Diagram showing unique hits for two different proteomics methods
}
