% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{Venn3.Fun}
\alias{Venn3.Fun}
\title{Three Method Comparison Venn Diagram}
\usage{
Venn3.Fun(data1, data2, data3, name1, name2, name3)
}
\arguments{
\item{data1}{unqiue hit csv output method 1}

\item{data2}{unique hit csv output method 2}

\item{name1}{method 1 name}

\item{name2}{method 2 name}
}
\value{
Venn Diagram
}
\description{
This function will make a Venn Diagram showing unique hits for three different proteomics methods
}
