% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{Venn4.Fun}
\alias{Venn4.Fun}
\title{Four Method Comparison Venn Diagram}
\usage{
Venn4.Fun(
  data1,
  data2,
  data3,
  data4,
  name1,
  name2,
  name3,
  name4,
  filename = "Four_Comparison_Venn_Diagram.PNG",
  imagetype = "png",
  textcolor = "black",
  outlinetype = "dotted",
  areafill = c("#6495ed", "#bf3eff", "#ffffba", "#90ee90"),
  alpha = c(0.3, 0.3, 0.3, 0.3),
  arealabelsize = 0.5,
  categorynamesize = 0.5,
  categorycolor = "black"
)
}
\arguments{
\item{data1}{Unqiue Hit .csv Output method 1}

\item{data2}{Unique Hit .csv Output method 2}

\item{data3}{Unique Hit .csv Output method 3}

\item{data4}{Unique Hit .csv Output method 4}

\item{name1}{Method 1 Name}

\item{name2}{Method 2 Name}

\item{name3}{Method 3 Name}

\item{name4}{Method 4 Name}

\item{filename}{Filename for Output (default = 'Four_Comparison_Venn_Diagram')}

\item{imagetype}{Image Output Extension (default = png)}

\item{textcolor}{Color of Circle's Circumference (default = black)}

\item{outlinetype}{Circle Outline Dash pattern (default = dotted)}

\item{areafill}{Color of Circles (default = c("#6495ed", "#bf3eff", "#ffffba", "#90ee90"))}

\item{alpha}{Transparency of Circles (default = 0.3)}

\item{arealabelsize}{Size of Area Labels (default = 0.5)}

\item{categorynamesize}{Size of Category Names (default = 0.5)}

\item{categorycolor}{Color of Category Names (default = black)}
}
\value{
Venn Diagram
}
\description{
This function will make a Venn Diagram showing unique hits for three different proteomics methods
}
