% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{ExpressionLevel.Fun}
\alias{ExpressionLevel.Fun}
\title{Load Expression data}
\usage{
ExpressionLevel.Fun(Expression_data, SD_cutoff, p_cutoff)
}
\arguments{
\item{Expression_data}{A data file}

\item{SD_cutoff}{A number}

\item{p_cutoff}{A number}
}
\value{
Hit list and volcano plot
}
\description{
This function will take expression data directly from Proteome Discoverer and generate a hit list
and Volcano Plot
}
