% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{OnePotSPROX.Fun}
\alias{OnePotSPROX.Fun}
\title{Load Expression data and SPROX data}
\usage{
OnePotSPROX.Fun(Expression_data, SPROX_Raw, SD_cutoff, p_cutoff)
}
\arguments{
\item{infile}{Path to the expression input file, SPROX data, SD_cutoff, and p_cutoff}
}
\value{
Hit list, volcano plot
}
\description{
This function will take expression data and SPROX data directly from Proteome Discoverer and generate a hit list
and Volcano Plot
}
