% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{OnePotTPP.Fun}
\alias{OnePotTPP.Fun}
\title{OnePotTPP Analysis}
\usage{
OnePotTPP.Fun(
  Expression_data,
  TPP_Raw,
  SD_cutoff,
  p_cutoff,
  TMTplex = 10,
  plot = TRUE,
  xlab = "Z Score",
  ylab = "- log10 (p value)",
  legendlabels = c("Not Significant", "Significant"),
  labelcolor = c("grey", "red"),
  alpha = 0.5
)
}
\arguments{
\item{Expression_data}{Raw ProteomeDiscoverer (v. 2.3) data}

\item{SD_cutoff}{A positive integer (default = 2)}

\item{p_cutoff}{A positive integer between 0 and 1 (default = 0.05)}

\item{TMTplex}{8, 10, 16 TMTplex (default = 10)}

\item{plot}{Display volcano plot (default = TRUE)}

\item{xlab}{X-axis Label (default = "Z Score")}

\item{ylab}{Y-axis Label (default = "- log10 (p value)")}

\item{legendlabels}{Name of groups in figure legend (default = c("Not Significant", "Significant))}

\item{labelcolor}{Color of Not significant and Significant data points (default = c("grey", "red"))}

\item{alpha}{Transparency of plot (default = 0.5)}

\item{TPP_data}{Raw OnePotTPP (TMT10plex) ProteomeDiscoverer (v. 2.3) data}
}
\value{
Volcano plot, Hit list, Total proteins assayed, and Unique protein hits
}
\description{
This function will take expression data and OnePotTPP (TMT10plex) data directly from ProteomeDiscoverer (v. 2.3) and generate a volcano plot and hit list
based on user-defined SD and p-values
}
