% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{STEPP.Fun}
\alias{STEPP.Fun}
\title{Load Expression data and STEPP data}
\usage{
STEPP.Fun(Expression_data, STEPP_Raw, SD_cutoff, p_cutoff)
}
\arguments{
\item{infile}{Path to the expression input file, STEPP data, SD_cutoff, and p_cutoff}
}
\value{
Hit list, volcano plot
}
\description{
This function will take expression data and STEPP data directly from Proteome Discoverer and generate a hit list
and Volcano Plot
}
