% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Proteomics_Functions.R
\name{Venn4.Fun}
\alias{Venn4.Fun}
\title{Four Method Comparison Venn Diagram}
\usage{
Venn4.Fun(data1, data2, data3, data4, name1, name2, name3, name4)
}
\arguments{
\item{data1}{Unqiue Hit .csv Output from method 1}

\item{data2}{Unique Hit .csv Output method 2}

\item{data3}{Unique Hit .csv Output method 3}

\item{data4}{Unique Hit .csv Output method 4}

\item{name1}{Method 1 name}

\item{name2}{Method 2 name}

\item{name3}{Method 3 name}

\item{name4}{Method 4 name}
}
\value{
Venn Diagram
}
\description{
This function will make a Venn Diagram showing unique hits for three different proteomics methods
}
